renamed styling to remove styles issue with git
