+++ web/src/Pages/TopicPage/BreadCrumbs/Breadcrumbs.jsx
@@ -1,7 +1,7 @@
 /* eslint-disable react/button-has-type */
 import { Link } from "react-router-dom";
 import { React } from "react";
-import "../../../Styles/TopicBreadcrumbs/TopicBreadcrumbs.css";
+import "../../../Styling/TopicBreadcrumbs/TopicBreadcrumbs.css";
 
 export default function Breadcrumbs({ grandParent, parent }) {
   return (
