+++ web/src/Pages/TopicPage/TopicCard/TopicCard.js
@@ -1,5 +1,5 @@
 import React from "react";
-import "../../../Styles/TopicCard/TopicCard.css";
+import "../../../Styling/TopicCard/TopicCard.css";
 
 export default function TopicCard({ topic, audience }) {
   const description = topic.descriptions?.find?.(({ audience: a }) => audience === a)?.short;
