+++ web/src/Styling/Tailwind/tailwind.css
@@ -1,0 +1,3 @@
+@tailwind base;
+@tailwind components;
+@tailwind utilities;/*# sourceMappingURL=tailwind.css.map */
