+++ web/src/Styling/Tailwind/tailwind.scss
@@ -1,0 +1,3 @@
+@tailwind base;
+@tailwind components;
+@tailwind utilities;
