+++ web/src/Styling/TopicCard/_variables.scss
